Reworking Projects component from scratch...
The previous version of the projects component was a mess.
I wanted to start over from scratch and rely on events instead
of react state. Also, the grid layout was more that was needed
to get the job done. I ended up using flexbox to layout child components
instead of grid, since the grid was always changing with the width of
the page.